feat(bet4-admin): extract admin operator tools into a scoped ToolPack (EP-8DC217EB BET-4)
Process-Spine-Decision: BET-4 drain-to-zero (plan §4) — behavior-identical pack extraction; gating metadata (requiredCapability/sideEffect tiers) preserved verbatim; runtime-gate test green.
BI: BI-2B7EE073
